Move typeahead plugin to shared-react
