+++ packages/perf-react/package.json
@@ -21,8 +21,7 @@
     "@lexical/html": "^0.17.1",
     "@lexical/react": "^0.17.1",
     "@lexical/selection": "^0.17.1",
-    "@lexical/utils": "^0.17.1",
-    "@floating-ui/dom": "^1.6.10"
+    "@lexical/utils": "^0.17.1"
   },
   "devDependencies": {
     "@types/react": "^18.3.1",
