+++ packages/shared-react/package.json
@@ -18,7 +18,8 @@
   "dependencies": {
     "@lexical/react": "^0.17.1",
     "@lexical/utils": "^0.17.1",
-    "lexical": "^0.17.1"
+    "lexical": "^0.17.1",
+    "@floating-ui/dom": "^1.6.10"
   },
   "devDependencies": {
     "@types/react": "^18.3.1",
