+++ packages/shared-react/plugins/FloatingBox/FloatingBox.tsx
