+++ packages/shared-react/plugins/FloatingBox/usePointerInteractions.tsx
