+++ packages/shared-react/plugins/ScriptureReferencePlugin.tsx
