+++ packages/shared-react/plugins/Typeahead/TypeaheadPlugin.tsx
@@ -1,8 +1,9 @@
 import TypeaheadFloatingBox from "./TypeaheadFLoatingBox";
-import AutocompleteMenu from "./AutocompleteMenu";
-import { AutoCompleteItem } from "./useAutocompleteItems";
+import AutocompleteMenu from "../Autocomplete/AutocompleteMenu";
+import { AutoCompleteItem } from "../Autocomplete/useAutocompleteItems";
 
 type TypeaheadPluginProps = {
+  /** the string that will trigger the floatingMenu */
   trigger: string;
   items?: AutoCompleteItem[];
 };
@@ -11,6 +12,14 @@ const filterFunction = (item: AutoCompleteItem, phrase: string): boolean =>
   item.label?.toLowerCase().includes(phrase.toLowerCase()) ||
   item.name.toLowerCase().includes(phrase.toLowerCase());
 
+/**
+ * A plugin that renders an autocomplete floating menu when the user triggers it
+ * by typing in a trigger character or phrase
+ *
+ * @param trigger the string that will trigger the typeahead menu when typed into the editor
+ * @param items an array of option items to be placed inside the menu
+ * @returns
+ */
 export default function TypeaheadPlugin({ trigger, items }: TypeaheadPluginProps) {
   return (
     <TypeaheadFloatingBox trigger={trigger}>
